refactor(auth): 이메일 인증 및 비밀번호 재설정 로직 개선

+++ src/main/java/kr/co/cloudStudy/auth/service/impl/EmailCodeServiceImpl.java
@@ -25,10 +25,15 @@ public class EmailCodeServiceImpl implements EmailCodeService {
 	
 	@Override
 	public void sendEmailCode(String email) {
+		
+		if (emailCodeRedisRepository.isSendBlocked(email)) {
+			throw new IllegalArgumentException("인증번호는 잠시 후 다시 요청해주세요.");
+		}
 		String code = createRandomCode();
 		
 		emailCodeRedisRepository.saveEmailCode(email, code);
 		emailCodeRedisRepository.deleteVerified(email);
+		emailCodeRedisRepository.saveSendBlock(email);
 		mailService.sendEmailCode(email, code);
 		
 	}
@@ -55,9 +60,11 @@ public class EmailCodeServiceImpl implements EmailCodeService {
 		Employee employee = employeeRepository.findByEmail(email)
 				.orElseThrow(() -> new IllegalArgumentException("해당 이메일의 사용자를 찾을 수 없습니다."));
 		
-		if (!emailCodeRedisRepository.isVerifired(email)) {
+		if (!emailCodeRedisRepository.isVerified(email)) {
 			throw new IllegalArgumentException("이메일 인증이 완료되지 않았습니다.");
 		}
+		
+		validatePassword(newPassword);
 		
 //		TODO: 나중에 반드시 복구 (비밀번호 암호화 적용)
 //		String encodedPassword = passwordEncoder.encode(newPassword);
@@ -72,6 +79,14 @@ public class EmailCodeServiceImpl implements EmailCodeService {
 		SecureRandom secureRandom = new SecureRandom();
 		int number = 100000 + secureRandom.nextInt(900000);
 		return String.valueOf(number);
+	}
+	
+	private void validatePassword(String password) {
+	    String regex = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*#?&]).{8,}$";
+
+	    if (!password.matches(regex)) {
+	        throw new IllegalArgumentException("비밀번호는 8자 이상이며 영문, 숫자, 특수문자를 포함해야 합니다.");
+	    }
 	}
 }
